requirements.txt
torch==2.6.0
torchvision==0.21.0
facenet-pytorch
numpy==1.26.4
matplotlib==3.7.2
scikit-learn==1.2.2
albumentations==2.0.5
Pillow